requirements.txt
PyYAML~=6.0
ipfshttpclient==0.6.1
pinatapy-vourhey==0.1.3
rospy~=1.15.13
requests~=2.26.0
rospkg
qrcode
Pillow==8.3.2
qrcode==6.1